Allows for JLabel in JMenu
